refactor: extract model classes into shared project
