+++ gradle.properties
@@ -4,3 +4,5 @@ logbackVersion=1.2.6
 
 org.gradle.caching=true
 kotlin.code.style=official
+kotlin.js.compiler=ir
+kotlin.mpp.stability.nowarn=true
