+++ server/src/main/kotlin/com/larsreimann/api_editor/server/Routing.kt
@@ -1,6 +1,6 @@
 package com.larsreimann.api_editor.server
 
-import com.larsreimann.api_editor.server.data.AnnotatedPythonPackage
+import com.larsreimann.api_editor.shared.model.AnnotatedPythonPackage
 import io.ktor.application.Application
 import io.ktor.application.call
 import io.ktor.application.install
