+++ server/src/test/kotlin/com/larsreimann/api_editor/server/ApplicationTest.kt
@@ -1,30 +1,30 @@
 package com.larsreimann.api_editor.server
 
-import com.larsreimann.api_editor.server.data.AnnotatedPythonClass
-import com.larsreimann.api_editor.server.data.AnnotatedPythonFunction
-import com.larsreimann.api_editor.server.data.AnnotatedPythonModule
-import com.larsreimann.api_editor.server.data.AnnotatedPythonPackage
-import com.larsreimann.api_editor.server.data.AnnotatedPythonParameter
-import com.larsreimann.api_editor.server.data.AnnotatedPythonResult
-import com.larsreimann.api_editor.server.data.AttributeAnnotation
-import com.larsreimann.api_editor.server.data.BoundaryAnnotation
-import com.larsreimann.api_editor.server.data.CalledAfterAnnotation
-import com.larsreimann.api_editor.server.data.ComparisonOperator
-import com.larsreimann.api_editor.server.data.ConstantAnnotation
-import com.larsreimann.api_editor.server.data.DefaultBoolean
-import com.larsreimann.api_editor.server.data.DefaultNumber
-import com.larsreimann.api_editor.server.data.DefaultString
-import com.larsreimann.api_editor.server.data.EnumAnnotation
-import com.larsreimann.api_editor.server.data.EnumPair
-import com.larsreimann.api_editor.server.data.GroupAnnotation
-import com.larsreimann.api_editor.server.data.MoveAnnotation
-import com.larsreimann.api_editor.server.data.OptionalAnnotation
-import com.larsreimann.api_editor.server.data.PythonFromImport
-import com.larsreimann.api_editor.server.data.PythonImport
-import com.larsreimann.api_editor.server.data.PythonParameterAssignment
-import com.larsreimann.api_editor.server.data.RenameAnnotation
-import com.larsreimann.api_editor.server.data.RequiredAnnotation
-import com.larsreimann.api_editor.server.data.UnusedAnnotation
+import com.larsreimann.api_editor.shared.model.AnnotatedPythonClass
+import com.larsreimann.api_editor.shared.model.AnnotatedPythonFunction
+import com.larsreimann.api_editor.shared.model.AnnotatedPythonModule
+import com.larsreimann.api_editor.shared.model.AnnotatedPythonPackage
+import com.larsreimann.api_editor.shared.model.AnnotatedPythonParameter
+import com.larsreimann.api_editor.shared.model.AnnotatedPythonResult
+import com.larsreimann.api_editor.shared.model.AttributeAnnotation
+import com.larsreimann.api_editor.shared.model.BoundaryAnnotation
+import com.larsreimann.api_editor.shared.model.CalledAfterAnnotation
+import com.larsreimann.api_editor.shared.model.ComparisonOperator
+import com.larsreimann.api_editor.shared.model.ConstantAnnotation
+import com.larsreimann.api_editor.shared.model.DefaultBoolean
+import com.larsreimann.api_editor.shared.model.DefaultNumber
+import com.larsreimann.api_editor.shared.model.DefaultString
+import com.larsreimann.api_editor.shared.model.EnumAnnotation
+import com.larsreimann.api_editor.shared.model.EnumPair
+import com.larsreimann.api_editor.shared.model.GroupAnnotation
+import com.larsreimann.api_editor.shared.model.MoveAnnotation
+import com.larsreimann.api_editor.shared.model.OptionalAnnotation
+import com.larsreimann.api_editor.shared.model.PythonFromImport
+import com.larsreimann.api_editor.shared.model.PythonImport
+import com.larsreimann.api_editor.shared.model.PythonParameterAssignment
+import com.larsreimann.api_editor.shared.model.RenameAnnotation
+import com.larsreimann.api_editor.shared.model.RequiredAnnotation
+import com.larsreimann.api_editor.shared.model.UnusedAnnotation
 import io.ktor.http.ContentType
 import io.ktor.http.HttpHeaders
 import io.ktor.http.HttpMethod
