+++ settings.gradle.kts
@@ -2,5 +2,6 @@ rootProject.name = "com.larsreimann.api_editor"
 
 include(
     "client",
-    "server"
+    "server",
+    "shared"
 )
